fix parameter types in the faad calls
libfaad wants uint32_t pointers.  Passing a long pointer is bugged on
amd64.

git-svn-id: https://svn.musicpd.org/mpd/trunk@7289 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/inputPlugins/aac_plugin.c
@@ -234,7 +234,7 @@ static float getAacFloatTotalTime(char *file)
 	size_t fileread, tagsize;
 	faacDecHandle decoder;
 	faacDecConfigurationPtr config;
-	unsigned long sampleRate;
+	uint32_t sampleRate;
 	unsigned char channels;
 	InputStream inStream;
 	long bread;
@@ -290,7 +290,7 @@ static int aac_decode(OutputBuffer * cb, DecoderControl * dc, char *path)
 	faacDecFrameInfo frameInfo;
 	faacDecConfigurationPtr config;
 	long bread;
-	unsigned long sampleRate;
+	uint32_t sampleRate;
 	unsigned char channels;
 	int eof = 0;
 	unsigned int sampleCount;
